+++ test/setup.js
@@ -42,6 +42,7 @@ global.trace = false
 includeJavascript('background/consts.js')
 includeJavascript('background/storage.js')
 includeJavascript('background/utils.js')
+includeJavascript('background/resource.js')
 includeJavascript('background/startup.js')
 includeJavascript('background/great_buildings.js')
 includeJavascript('background/city_map.js')